MeshCore/requirements.txt
meshcore
aiohttp
requests
pypubsub
# standard library-like modules
